[improve][client] Don't print info logs for each schema loaded by client
### Motivation

Currently when a client loads schema from brokers, it will print the
whole schema as info logs. The broker's built-in client has the same
issue so schemas will be printed in broker logs, which could be very
long (hundreds of lines) and might expose user info.

### Modifications

Reduce the log level to debug in `MultiVersionAvroReader`.