[fix][broker] fix the log format error in the AdminResource
## Motivation
The log message in `AdminResource` was missing a parameter. The `topicName` was not being included in the success log message when creating partitioned topics for remote clusters.

This results the following errors:
```
ERROR Unable to format msg: [{}] Successfully created partitioned for topic {} for the remote clusters java.lang.IllegalArgumentException: found 2 argument placeholders, but provided 1 for pattern `[{}] Successfully created partitioned for topic {} for the remote clusters`
```

## Modifications
Corrected the log message by adding the missing `topicName` parameter to the success log entry in the `internalCreatePartitionedTopicToReplicatedClustersInBackground` method.