[improve][broker] Clean up config changes without a PIP since the 4.0 LTS
- Remove `managedLedgerOffloadReadThreads` from https://github.com/apache/pulsar/pull/24025
- Remove `managedLedgerMaxReadsInFlightPermitsAcquireTimeoutMillis` and
  `managedLedgerMaxReadsInFlightPermitsAcquireQueueSize` from https://github.com/apache/pulsar/pull/23901
- Remove `managedLedgerCursorResetLedgerCloseTimestampMaxClockSkewMillis` from https://github.com/apache/pulsar/pull/22792
The configs above only increase the complexity and are hard to configure.

Add more comments to `managedLedgerPersistIndividualAckAsLongArray` from https://github.com/apache/pulsar/pull/23759.
This config was added to keep the compatibility from 3.x or earlier so
it has value to retain. 3.x users must configure it with false when
upgrading to 4.0.