[fix][test] Simplify BetweenTestClassesListenerAdapter and fix issue with BeforeTest/AfterTest annotations (#24304)
### Motivation

It wasn't clear how test class changes should be detected with TestNG. 
The solution was recently revisited in #24258. However problems remain after adding that solution. BeforeTest/AfterTest annotation usage won't work and cause false reports for detected leaks.

It turns out that it's possible to detect a test change with ITestListener's `onFinish` callback. In TestNG, it's possible to group multiple test classes into a single test, but this is only when using TestNG's XML configuration for test suites. maven-surefire-plugin will create a separate TestNG test for each test class, so that's why it's the correct solution for BetweenTestClassesListenerAdapter.

### Modifications

- Remove previous logic to detect test class change and instead rely on ITestListener's `onFinish` callback
- Adapt the current listener implementations to the new base class